Move body listening handling to the decorator
